back/pyproject.toml
[tool.poetry]
name = "djangodiplom"
version = "0.1.0"
description = ""
authors = ["Veronika <verasos13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
django-easy-select2 = "^1.5.8"
dj-database-url = "^0.5.0"
dbt-postgres = "^1.1.0"
networkx = "^2.8"
numpy = "^1.22.3"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
